added gauge

+++ upload.sh
@@ -1,32 +1,42 @@
 #!/bin/sh
-(
+for panel in trend gauge; do
+  (
 cat << 'EOF'
-simple-performance-single-server-all-time single-server-all-time.json
-simple-performance-single-server-6-months single-server-6-months.json
-simple-performance-single-server-6-months single-server-6-months.json
-simple-performance-single-server-1-month  single-server-1-month.json
-simple-performance-single-server-2-weeks  single-server-2-weeks.json
-simple-performance-single-server-2-days   single-server-2-days.json
+simple-performance-single-server-all-time single-server-all-time
+simple-performance-single-server-6-months single-server-6-months
+simple-performance-single-server-6-months single-server-6-months
+simple-performance-single-server-1-month  single-server-1-month
+simple-performance-single-server-2-weeks  single-server-2-weeks
+simple-performance-single-server-2-days   single-server-2-days
 EOF
-) | while read a b; do
-  curl "https://grafana.arangodb.biz/api/snapshots/$a" \
-    -H 'accept: application/json, text/plain, */*' \
-    -H "x-grafana-org-id: ${GRAFANA_ORG_ID}" \
-    -H 'content-type: application/json' \
-    -H "Authorization: Bearer ${GRAFANA_API_KEY}" \
-    -k \
-    -X DELETE
-  echo
+  ) | while read a b; do
+    file=${b}-${panel}.json
 
-  curl 'https://grafana.arangodb.biz/api/snapshots' \
-    -H 'accept: application/json, text/plain, */*' \
-    -H "x-grafana-org-id: ${GRAFANA_ORG_ID}" \
-    -H 'content-type: application/json' \
-    -H "Authorization: Bearer ${GRAFANA_API_KEY}" \
-    -k \
-    --data-binary @$b \
-    --compressed
-  echo
+    if test -f $file; then
+      echo "Uploading $file"
+
+      curl "https://grafana.arangodb.biz/api/snapshots/$a-${panel}" \
+        -H 'accept: application/json, text/plain, */*' \
+        -H "x-grafana-org-id: ${GRAFANA_ORG_ID}" \
+        -H 'content-type: application/json' \
+        -H "Authorization: Bearer ${GRAFANA_API_KEY}" \
+        -k \
+        -X DELETE
+      echo
+
+      curl 'https://grafana.arangodb.biz/api/snapshots' \
+        -H 'accept: application/json, text/plain, */*' \
+        -H "x-grafana-org-id: ${GRAFANA_ORG_ID}" \
+        -H 'content-type: application/json' \
+        -H "Authorization: Bearer ${GRAFANA_API_KEY}" \
+        -k \
+        --data-binary @${file} \
+        --compressed
+      echo
+    else
+      echo "ERROR: $file missing"
+    fi
+  done
 done
 
 curl 'https://grafana.arangodb.biz/api/snapshots/simple-performance-cluster' \
